Move some stuff to extension package
