+++ launcher/gradle-plugin/src/main/kotlin/org/sourcegrade/jagr/gradle/extension/JagrExtension.kt
@@ -1,0 +1,32 @@
+/*
+ *   Jagr - SourceGrade.org
+ *   Copyright (C) 2021-2022 Alexander Staeding
+ *   Copyright (C) 2021-2022 Contributors
+ *
+ *     This program is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU Affero General Public License as published by
+ *     the Free Software Foundation, either version 3 of the License, or
+ *     (at your option) any later version.
+ *
+ *     This program is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *     GNU Affero General Public License for more details.
+ *
+ *     You should have received a copy of the GNU Affero General Public License
+ *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package org.sourcegrade.jagr.gradle.extension
+
+import org.gradle.api.NamedDomainObjectContainer
+import org.gradle.api.plugins.ExtensionAware
+import org.gradle.api.provider.Property
+
+abstract class JagrExtension : ExtensionAware {
+    abstract val assignmentId: Property<String>
+    abstract val graders: NamedDomainObjectContainer<GraderConfiguration>
+    abstract val submissions: NamedDomainObjectContainer<SubmissionConfiguration>
+    fun graders(configure: NamedDomainObjectContainer<GraderConfiguration>.() -> Unit) = graders.configure()
+    fun submissions(configure: NamedDomainObjectContainer<SubmissionConfiguration>.() -> Unit) = submissions.configure()
+}
