+++ launcher/gradle-plugin/src/main/kotlin/org/sourcegrade/jagr/gradle/task/JagrTaskFactory.kt
@@ -2,7 +2,7 @@ package org.sourcegrade.jagr.gradle.task
 
 import org.gradle.api.Project
 import org.gradle.api.Task
-import org.sourcegrade.jagr.gradle.AbstractConfiguration
+import org.sourcegrade.jagr.gradle.extension.AbstractConfiguration
 
 internal interface JagrTaskFactory<T : Task, C : AbstractConfiguration> {
     fun determineTaskName(name: String): String
