+++ launcher/gradle-plugin/src/main/kotlin/org/sourcegrade/jagr/gradle/task/submission/SubmissionTask.kt
@@ -25,8 +25,8 @@ import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.TaskProvider
 import org.gradle.kotlin.dsl.getByType
 import org.gradle.kotlin.dsl.register
-import org.sourcegrade.jagr.gradle.JagrExtension
-import org.sourcegrade.jagr.gradle.SubmissionConfiguration
+import org.sourcegrade.jagr.gradle.extension.JagrExtension
+import org.sourcegrade.jagr.gradle.extension.SubmissionConfiguration
 import org.sourcegrade.jagr.gradle.task.JagrTaskFactory
 import org.sourcegrade.jagr.gradle.task.TargetAssignmentTask
 import org.sourcegrade.jagr.gradle.task.TargetSourceSetsTask
